chore: Make cdktf-project file a bit clearer to read
